Finished Project.
Project is finished for now, I've completed everything that the
project asked for, plus a few additional sections. I ended up making
this something that I could use in the future for generating a CV. Some
UI improvements that I would consider adding in the future is the
ability to add more sections, or remove unnecessary sections. On to the
next!